Adjust logos and stepper component
