+++ tmt-app/src/Components/Receipt.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Container } from "@mui/material";
-import QuizStepper from "./Stepper";
+import QuizStepper from "./QuizStepper";
 
 function Receipt() {
   return (
